refactor: JournalTr, JournalOutput, op using only revm crate
