+++ crates/context/interface/src/lib.rs
@@ -16,5 +16,5 @@ pub use block::Block;
 pub use cfg::{Cfg, CreateScheme, TransactTo};
 pub use context::ContextTr;
 pub use database_interface::{DBErrorMarker, Database};
-pub use journaled_state::Journal;
+pub use journaled_state::JournalTr;
 pub use transaction::{Transaction, TransactionType};
