+++ crates/context/src/lib.rs
@@ -11,14 +11,12 @@ pub mod block;
 pub mod cfg;
 pub mod context;
 pub mod evm;
-mod journal_init;
 pub mod journaled_state;
 pub mod tx;
 
 pub use block::BlockEnv;
 pub use cfg::{Cfg, CfgEnv};
 pub use context::*;
-pub use journal_init::JournalInit;
 pub use journaled_state::*;
 pub use tx::TxEnv;
 pub mod setters;
