+++ crates/optimism/src/api/exec.rs
@@ -2,19 +2,17 @@ use crate::{
     evm::OpEvm, handler::OpHandler, transaction::OpTxTr, L1BlockInfo, OpHaltReason, OpSpecId,
     OpTransactionError,
 };
-use inspector::{InspectCommitEvm, InspectEvm, Inspector, JournalExt};
-use precompile::Log;
 use revm::{
+    context::JournalOutput,
     context_interface::{
         result::{EVMError, ExecutionResult, ResultAndState},
-        Block, Cfg, ContextTr, Database, Journal,
+        Block, Cfg, ContextTr, Database, JournalTr,
     },
     handler::{instructions::EthInstructions, EthFrame, EvmTr, Handler, PrecompileProvider},
+    inspector::{InspectCommitEvm, InspectEvm, Inspector, JournalExt},
     interpreter::{interpreter::EthInterpreter, InterpreterResult},
-    state::EvmState,
     Context, DatabaseCommit, ExecuteCommitEvm, ExecuteEvm,
 };
-use std::vec::Vec;
 
 impl<BLOCK, TX, CFG, DB, JOURNAL, INSP, PRECOMPILE> ExecuteEvm
     for OpEvm<
@@ -28,7 +26,7 @@ where
     TX: OpTxTr,
     CFG: Cfg<Spec = OpSpecId>,
     DB: Database,
-    JOURNAL: Journal<Database = DB, FinalOutput = (EvmState, Vec<Log>)>,
+    JOURNAL: JournalTr<Database = DB, FinalOutput = JournalOutput>,
     PRECOMPILE: PrecompileProvider<
         Context = Context<BLOCK, TX, CFG, DB, JOURNAL, L1BlockInfo>,
         Output = InterpreterResult,
@@ -55,7 +53,7 @@ where
     TX: OpTxTr,
     CFG: Cfg<Spec = OpSpecId>,
     DB: Database + DatabaseCommit,
-    JOURNAL: Journal<Database = DB, FinalOutput = (EvmState, Vec<Log>)> + JournalExt,
+    JOURNAL: JournalTr<Database = DB, FinalOutput = JournalOutput> + JournalExt,
     PRECOMPILE: PrecompileProvider<
         Context = Context<BLOCK, TX, CFG, DB, JOURNAL, L1BlockInfo>,
         Output = InterpreterResult,
@@ -86,7 +84,7 @@ where
     TX: OpTxTr,
     CFG: Cfg<Spec = OpSpecId>,
     DB: Database,
-    JOURNAL: Journal<Database = DB, FinalOutput = (EvmState, Vec<Log>)> + JournalExt,
+    JOURNAL: JournalTr<Database = DB, FinalOutput = JournalOutput> + JournalExt,
     INSP: Inspector<Context<BLOCK, TX, CFG, DB, JOURNAL, L1BlockInfo>, EthInterpreter>,
     PRECOMPILE: PrecompileProvider<
         Context = Context<BLOCK, TX, CFG, DB, JOURNAL, L1BlockInfo>,
@@ -117,7 +115,7 @@ where
     TX: OpTxTr,
     CFG: Cfg<Spec = OpSpecId>,
     DB: Database + DatabaseCommit,
-    JOURNAL: Journal<Database = DB, FinalOutput = (EvmState, Vec<Log>)> + JournalExt,
+    JOURNAL: JournalTr<Database = DB, FinalOutput = JournalOutput> + JournalExt,
     INSP: Inspector<Context<BLOCK, TX, CFG, DB, JOURNAL, L1BlockInfo>, EthInterpreter>,
     PRECOMPILE: PrecompileProvider<
         Context = Context<BLOCK, TX, CFG, DB, JOURNAL, L1BlockInfo>,
