+++ crates/revm/Cargo.toml
@@ -65,8 +65,6 @@ alloydb = ["database/alloydb"]
 # Enables serde-json inside inspector crate
 serde-json = ["serde", "inspector/serde-json"]
 
-test-utils = []
-
 dev = [
     "memory_limit",
     "optional_balance_check",
@@ -80,9 +78,12 @@ optional_block_gas_limit = ["context/optional_block_gas_limit"]
 optional_eip3607 = ["context/optional_eip3607"]
 optional_no_base_fee = ["context/optional_no_base_fee"]
 
+# Precompiles features
+
 # See comments in `precompile`
 secp256k1 = ["precompile/secp256k1"]
 c-kzg = ["precompile/c-kzg"]
 # `kzg-rs` is not audited but useful for `no_std` environment, use it with causing and default to `c-kzg` if possible.
 kzg-rs = ["precompile/kzg-rs"]
 blst = ["precompile/blst"]
+secp256r1 = ["precompile/secp256r1"]
